Added pages to the DJ journey and change of routes.js to local copy inside DJ folder
