+++ app/views/default-judgments/2-defendant.html
@@ -40,7 +40,7 @@
                 </label>
               </div>
               <div class="govuk-radios__item">
-                <input class="govuk-radios__input" id="defendant3" name="defendant" type="radio" value="Both defendants">
+                <input class="govuk-radios__input" id="all" name="defendant" type="radio" value="Both defendants">
                 <label class="govuk-label govuk-radios__label" for="both defendants">
                   Both defendants
                 </label>
